pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "exoplanet-analysis-report"
version = "0.1.0"
description = "A Data Analysis Report for Exoplanet data"
readme = "README.md"
authors = [
    { name = "Shiki Ichitose", email = "149179099+ShikiIchitose@users.noreply.github.com" }
]
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "httpx",
    "pandas",
    "numpy",
    "pyarrow",
    "duckdb",
    "matplotlib",
    "markdown-it-py",
]

[project.scripts]
exo-analyze = "exoplanet_analysis_report.cli:main"

# ---- dev deps (uv) ----
[dependency-groups]
dev = [
  "ruff",
  "pytest",
  "pytest-socket",
]

# ---- dev deps (pip extras) ----
# pip users: `pip install -e ".[dev]"`
[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
  "pytest-socket",
]

# ---- setuptools (src-layout) ----
# Make package discovery explicit so `pip install -e .` is predictable.
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "I001"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "-ra",
  "--strict-markers",
  "--disable-socket",
]